refactor : 이벤트 큐이름 변경에 따른 코드 수정

+++ src/main/java/com/example/surveyapi/domain/statistic/application/event/ParticipationResponses.java
@@ -16,7 +16,7 @@ public record ParticipationResponses(
 ) {
 	public record Answer(
 		Long questionId,
-		List<Long> choiceIds,
+		List<Integer> choiceIds,
 		String responseText
 	) {}
 }
